+++ CHANGES.md
@@ -8,6 +8,7 @@ wavesurfer.js changelog
 - Fix microphone plugin example for Safari (#1377)
 - Minimap plugin: fix styling issues and add support for zooming (#1464)
 - Timeline plugin: add duration parameter handling (#1491)
+- Cursor plugin: add showTime option (#1143)
 - Fix: progress bar did not reach 100% when audio file is small (#1502)
 
 2.1.0 (29.09.2018)